rust: macros: decouple module parameters from generated constants
This adds a parameter `command_line_name` that allows specifying
an alternative name of a parameter for use on the command line:

    module! {
        type: ParamName,
        // ...
        params: {
            MODPARAM_FOO_BAR: bool {
                command_line_name: "foobar"
                default: true,
                permissions: 0o644,
                description: "test",
            },
        },
    }

In this example the parameter `foobar` is available on the Rust
end as `MODPARAM_FOO_BAR`.

The rationale is that for each parameter the `module!` macro
generates a constant with the same name as the parameter. This
can lead to surprising behavior when attempting to create a let
binding of the same name:

    module! {
        type: ParamName,
        // ...
        params: {
            foobar : bool {
                default: true,
                permissions: 0o644,
                description: "test",
            },
        },
    }

    fn clash() {
        let foobar = 42;
    }

This causes rustc to error out because `foobar` is treated as a
pattern on the left hand side of the let expression:

    10 | / module! {
    11 | |     type: ParamName,
    12 | |     name: "param_name",
    13 | |     author: "A. U. Thor",
    ...  |
    21 | |     },
    22 | | }
       | |_- constant defined here
    ...
    25 |       let foobar = 42;
       |           ^^^^^^   -- this expression has type `{integer}`
       |           |
       |           expected integer, found struct `__param_name_foobar`
       |           `foobar` is interpreted as a constant, not a new binding
       |           help: introduce a new binding instead: `other_foobar`

Essentially, having a parameter `foobar` prevents the creation of
a binding of the same name anywhere else in the module.

Signed-off-by: Philipp Gesang <phg@phi-gamma.net>