rust: macros: expose module parameters through MODPARAM module
Instead of creating a struct + constant pair for each parameter,
expose the parameters as functions in a submodule MODPARAM. For
example, the parameter `foobar` becomes available at the Rust end
as `MODPARAM::foobar()`.

The rationale is that currently the `module!` macro generates a
constant with the name of the parameter which can lead to
surprising behavior when attempting to create a let binding of
the same name:

    module! {
        type: ParamName,
        // ...
        params: {
            foobar : bool {
                default: true,
                permissions: 0o644,
                description: "test",
            },
        },
    }

    fn clash() {
        let foobar = 42;
    }

This causes rustc to error out because `foobar` is treated as a
pattern on the left hand side of the let expression:

    10 | / module! {
    11 | |     type: ParamName,
    12 | |     name: "param_name",
    13 | |     author: "A. U. Thor",
    ...  |
    21 | |     },
    22 | | }
       | |_- constant defined here
    ...
    25 |       let foobar = 42;
       |           ^^^^^^   -- this expression has type `{integer}`
       |           |
       |           expected integer, found struct `__param_name_foobar`
       |           `foobar` is interpreted as a constant, not a new binding
       |           help: introduce a new binding instead: `other_foobar`

Essentially, having a parameter `foobar` prevents the creation of
a binding of the same name anywhere else in the module.

Signed-off-by: Philipp Gesang <phg@phi-gamma.net>